@@ -1,49 +1,51 @@
 {
   "name": "wodin",
-  "version": "0.3.0",
-  "description": "Create web interfaces for odin",
-  "main": "dist/server.js",
+  "private": true,
+  "version": "1.0.0",
   "scripts": {
-    "build": "tsc",
-    "serve-hot": "node dist/server/server.js --hot-reload=true ../../config",
+    "build": "tsc --project tsconfig.node.json",
     "serve": "node dist/server/server.js ../../config",
-    "test": "jest -c jest.config.js",
-    "integration-test": "jest -c jest.config.integration.js",
-    "lint": "eslint src tests --ext .ts,.js",
-    "lint:fix": "eslint src tests --ext .ts,.js --fix",
+    "serve-hot": "node dist/server/server.js --hot-reload=true ../../config",
+    "dev": "nodemon src/server/server.ts ../../config",
+    "dev-hot": "nodemon src/server/server.ts --hot-reload=true ../../config",
+    "test:unit": "vitest -c ./vitest/vitest.unit.config.mjs",
+    "test:integration": "vitest -c ./vitest/vitest.integration.config.mjs run",
+    "coverage": "vitest -c ./vitest/vitest.unit.config.mjs run --coverage",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
     "genversion": "genversion --es6 --semi --double src/version.ts"
   },
-  "bin": "./dist/server.js",
+  "devDependencies": {
+    "@eslint/js": "^9.14.0",
+    "@types/compression": "^1.7.5",
+    "@types/docopt": "^0.6.37",
+    "@types/eslint__js": "^8.42.3",
+    "@types/express": "^5.0.0",
+    "@types/md5": "^2.3.5",
+    "@types/morgan": "^1.9.9",
+    "@types/mustache": "^4.2.5",
+    "@types/supertest": "^6.0.2",
+    "@vitest/coverage-istanbul": "^2.1.4",
+    "axios-mock-adapter": "^2.1.0",
+    "eslint": "^9.14.0",
+    "genversion": "^3.2.0",
+    "nodemon": "^3.1.7",
+    "supertest": "^7.0.0",
+    "ts-node": "^10.9.2",
+    "typescript": "~5.6.2",
+    "typescript-eslint": "^8.13.0"
+  },
   "dependencies": {
-    "@types/docopt": "^0.6.33",
-    "axios": "^0.27.2",
-    "body-parser": "^1.20.0",
-    "compression": "^1.7.4",
+    "axios": "^1.7.7",
+    "compression": "^1.7.5",
     "docopt": "^0.6.2",
-    "express": "^4.17.2",
-    "hbs": "^4.2.0",
-    "ioredis": "^5.2.2",
+    "express": "^4.21.1",
+    "ioredis": "^5.4.1",
     "md5": "^2.3.0",
     "morgan": "^1.10.0",
-    "uid": "^2.0.0",
-    "zoo-ids": "^2.0.0"
-  },
-  "devDependencies": {
-    "@types/axios": "^0.14.0",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^27.4.0",
-    "@types/md5": "^2.3.4",
-    "@types/node": "^17.0.12",
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
-    "axios-mock-adapter": "^1.20.0",
-    "eslint": "^8.8.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.4",
-    "genversion": "^3.1.1",
-    "jest": "^27.4.7",
-    "supertest": "^6.3.0",
-    "ts-jest": "^27.1.3",
-    "typescript": "^4.5.5"
+    "mustache": "^4.2.0",
+    "uid": "^2.0.2",
+    "vitest": "^2.1.4",
+    "zoo-ids": "^2.0.7"
   }
 }
