@@ -1,3 +1,3 @@
-const request = require("supertest");
+import request from "supertest";
 
 export const wodin = request("http://localhost:3000");
