@@ -1,2 +1,2 @@
 // Generated by genversion.
-export const version = "0.3.0";
+export const version = "1.0.0";
